SWI and get for Light is complete, still need to test
The SWI needs to be enabled

+++ sensors/sensor_api.h
@@ -19,7 +19,14 @@
 #include "inc/hw_memmap.h"
 #include <ti/drivers/UART.h>
 #include "motor/motor_api.h"
+
+// All the port macros are defined here
 #include "sensor_ports.h"
+
+
+#include "opt3001/opt3001.h"
+#include "opt3001/i2cOptDriver.h"
+
 
 
 // Initialises all the sensors:
@@ -27,7 +34,7 @@
 bool initSensors(uint8_t accel_threshold);
 
 // Read and filter light over I2C
-uint8_t getLight();
+float getLight();
 
 // Read and filter board and motor temperature sensors over UART
 uint8_t getBoardTemp();
